Without aufg bid

+++ src/main/java/solutions/dmitrikonnov/einstufungstest/domainlayer/ETAufgabenBogenDto.java
@@ -9,6 +9,6 @@ import java.util.List;
 public class ETAufgabenBogenDto implements Serializable {
 
     private static final long serialVersionUID = -1798070786993154676L;
-    private final Long aufgabenBogenId;
+    private final Integer aufgabenBogenId;
     private final List<ETAufgabeDto> aufgabenListe;
 }
